+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Multi-Form Goal block and shortcode are now available (#5307)
 -   Progress Bar block now features improved SSR previews (#5316)
 -   Multi-Form Goal block now supports "wide" alignment (#5315)
+-   Multi-Form Goal block now supports the theme's color palette (#5319)
 -   Multi-Form Goal block and shortcode appearance is now consistent (#5320)
 
 ### Changed